requirements.txt
# Web framework
Flask==2.3.6
Flask-Cors==3.0.13

# HTTP requests
requests==2.32.0

# Image processing
Pillow==10.0.0
pdf2image==1.16.3
opencv-python==4.8.1.78

# Machine learning / object detection
torch==2.2.0
torchvision==0.17.1
ultralytics==8.1.0

# Google Gemini API
google-generativeai==0.3.0

# Numerical / scientific computing
numpy==1.26.2
scipy==1.12.2

# Utilities / debugging
rich==13.6.0
